Ankur:5-Nov-2017 Added API for clients table GET - /v1/clients -> retrieve all clientSince GET - /v1/clients/id -> retrieve specific client DELETE - /v1/clients/id -> delete specific client PUT - /v1/clients/id -> update a client POST - /v1/clients/id -> create a client

+++ src/main/java/com/thinkhr/external/api/repositories/CompanyRepository.java
@@ -13,5 +13,5 @@ import com.thinkhr.external.api.db.entities.Company;
  *
  */
 
-public interface CompanyRepository extends CrudRepository<Company, Long> {
+public interface CompanyRepository extends CrudRepository<Company, Integer> {
 }
